Increase of precision of perspective parameters in UI (issue 1781) pp3 version change to 307

+++ rtengine/procparams.h
@@ -565,8 +565,8 @@ public:
 class PerspectiveParams {
 
     public:
-		int  horizontal;
-		int  vertical;
+		double  horizontal;
+		double  vertical;
 };
 
 /**
